requirements.txt
bcrypt==3.2.2
cachelib==0.9.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
Flask==2.1.3
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
flask-swagger-ui==4.11.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-sqlalchemy==0.28.1
mysql-connector==2.2.9
mysql-connector-python==8.0.30
oauthlib==2.1.0
packaging==21.3
pdfkit==1.0.0
protobuf==3.20.1
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
python-dotenv==0.20.0
requests==2.28.1
requests-oauthlib==1.1.0
six==1.16.0
SQLAlchemy==1.4.39
urllib3==1.26.12
Werkzeug==2.2.1
